+++ test/apps/v6-app/src/App.tsx
@@ -14,7 +14,7 @@ import * as React from 'react';
 import { Route, Routes, useNavigate } from 'react-router-dom';
 import { OktaAuth, toRelativeUrl } from '@okta/okta-auth-js';
 import { Security, LoginCallback } from '@okta/okta-react';
-import { SecureRoute } from './SecureRoute';
+import { AuthRequired } from '@okta/okta-react/react-router-6';
 import Home from './Home';
 import Protected from './Protected';
 import CustomLogin from './CustomLogin';
@@ -51,7 +51,7 @@ const App: React.FC<{
           <Route path='/login' element={<CustomLogin />} />
           <Route path='/widget-login'  element={<WidgetLogin baseUrl={baseUrl} />} />
           <Route path='/sessionToken-login' element={<SessionTokenLogin />} />
-          <Route path='/protected' element={<SecureRoute />}>
+          <Route path='/protected' element={<AuthRequired />}>
             <Route path='' element={<Protected />} />
           </Route>
           <Route path='/implicit/callback' element={<LoginCallback />} />